+++ src/components/select/Select.js
@@ -391,6 +391,7 @@ export class SelectComponent extends BaseComponent {
         containerOuter: 'choices form-group formio-choices',
         containerInner: 'form-control'
       },
+      itemComparer: _isEqual,
       placeholder: !!this.component.placeholder,
       placeholderValue: placeholderValue,
       searchPlaceholderValue: placeholderValue,
@@ -408,6 +409,7 @@ export class SelectComponent extends BaseComponent {
     // If a search field is provided, then add an event listener to update items on search.
     if (this.component.searchField) {
       this.addEventListener(input, 'search', (event) => this.triggerUpdate(event.detail.value));
+      this.addEventListener(input, 'stopSearch', () => this.triggerUpdate());
     }
 
     this.addEventListener(input, 'showDropdown', () => {
@@ -520,8 +522,10 @@ export class SelectComponent extends BaseComponent {
     if (this.choices) {
       // Now set the value.
       if (hasValue) {
-        this.choices.setChoices(this.selectOptions, 'value', 'label', true);
-        this.choices.setValueByChoice(_isArray(value) ? value : [value])
+        this.choices
+          .removeActiveItems()
+          .setChoices(this.selectOptions, 'value', 'label', true)
+          .setValueByChoice(_isArray(value) ? value : [value])
       }
       else if (hasPreviousValue) {
         this.choices.removeActiveItems();